[ALPS06336443] Revert "Merge android12-5.10 into alps-mp-s0.mp1"
This reverts commit 980e08c9a9741fcf32652570dbf6bd46d0e97ebd.
and revert other 3 chain changes: 8839d7e1c60 bde62c321729 535daf890f17

MTK-Commit-Id: 1071f55fd622b73df8d28ed98c9ed62c91f438b4

Feature: Kernel SI Operation
CR-Id: ALPS06336443

+++ arch/hexagon/include/asm/timex.h
@@ -8,7 +8,6 @@
 
 #include <asm-generic/timex.h>
 #include <asm/timer-regs.h>
-#include <asm/hexagon_vm.h>
 
 /* Using TCX0 as our clock.  CLOCK_TICK_RATE scheduled to be removed. */
 #define CLOCK_TICK_RATE              TCX0_CLK_RATE
@@ -17,7 +16,7 @@
 
 static inline int read_current_timer(unsigned long *timer_val)
 {
-	*timer_val = __vmgettime();
+	*timer_val = (unsigned long) __vmgettime();
 	return 0;
 }
